requirements.txt
selenium
python-dotenv
tinydb
webdriver_manager
scipy